% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{action_task}
\alias{action_task}
\title{Action a task}
\usage{
action_task(task_uuid, message = NULL, use_proxy = FALSE)
}
\arguments{
\item{task_uuid}{UUID of task to action}

\item{message}{Optionally, a character string to attach as a message to the
action.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
API response (invisibly)
}
\description{
Action a task
}
\details{
Actioning a task will change the assignee status from "incomplete"
to "complete". If all assignees have completed the task, the task status will
also change to "complete".

It is not possible to action a task if:
\itemize{
\item the task has already been completed or cancelled,
\item the task isn't assigned to the current authenticated user.
}
}
