% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{add_participants}
\alias{add_participants}
\title{Add participant(s) to a workspace}
\usage{
add_participants(
  workspace_uuid,
  emails,
  message = NULL,
  permissions = NULL,
  member_visibility = c("standard", "bcc"),
  send_email_invite = TRUE,
  use_proxy = FALSE
)
}
\arguments{
\item{workspace_uuid}{UUID of workspace.}

\item{emails}{Character vector of email addresses to send invites to.}

\item{message}{Optionally, a message to include in email invite.}

\item{permissions}{Optionally, a character vector of permissions to give
invited participants.

Valid permissions are: "Download", "CreateDocument",
"CreateFolder", "Edit", "Delete", "EditOnline", "Inviter", "Commenter" and
"ManageWorkspace".

All members are given permission to preview documents.}

\item{member_visibility}{Either "standard" (default) to make new participants
visible to all other workspace members, or "bcc" to hide participants.}

\item{send_email_invite}{Default \code{TRUE} to send an email invite for
participants to join the workspace. If \code{FALSE}, no email will be sent.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
API response (invisibly)
}
\description{
Add participant(s) to a workspace
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Participants/addParticipants}{API documentation}.
}
