% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{asset_info}
\alias{asset_info}
\title{Get asset information}
\usage{
asset_info(asset_uuid, use_proxy = FALSE)
}
\arguments{
\item{asset_uuid}{UUID of asset}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Named list containing: uuid, name, type, extension, description.
}
\description{
Get asset information
}
