% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets}
\alias{assets}
\title{Get data frame of assets in workspace}
\usage{
assets(
  workspace_uuid,
  type = list(),
  page = NULL,
  size = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{workspace_uuid}{UUID of workspace}

\item{type}{List of asset types to return. Default returns all types;
"document", "folder" and "link".

List must be empty (default, returns all asset types), or length 1 (e.g.
\code{list("document")}). This is a temporary measure while a bug in the
underlying API is outstanding (see
\href{https://github.com/ScotGovAnalysis/objr/issues/53}{objr#53}).}

\item{page}{Page number of responses to return (0..N). Default is 0.}

\item{size}{Number of results to be returned per page. Default is 20.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
A tibble
}
\description{
Get data frame of assets in workspace
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/getAssets}{API documentation}.
}
\seealso{
Other Asset functions: 
\code{\link{asset_info}()},
\code{\link{create_folder}()},
\code{\link{delete_asset}()},
\code{\link{rename_asset}()}
}
\concept{Asset functions}
