% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{cancel_task}
\alias{cancel_task}
\title{Cancel a task}
\usage{
cancel_task(task_uuid, message = NULL, use_proxy = FALSE)
}
\arguments{
\item{task_uuid}{UUID of task to cancel}

\item{message}{Optionally, a character string to attach as a message to the
cancellation.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
API response (invisibly)
}
\description{
Cancel a task
}
\details{
Cancelling a task will change its status from "open" to "cancelled".
}
