% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_valid}
\alias{check_valid}
\title{Helper function for validating input value}
\usage{
check_valid(value, warn = FALSE)
}
\arguments{
\item{value}{Value to verify}

\item{warn}{Logical. Indicates whether to produce warning if value is invalid}
}
\value{
Validated value
}
\description{
Helper function for validating input value
}
