% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveR_auth.R
\name{credential_helper}
\alias{credential_helper}
\title{Helper function for getting valid credentials}
\usage{
credential_helper(type = c("usr", "pwd"), value = NULL)
}
\arguments{
\item{type}{Either 'usr' (username) or 'pwd' (password).}

\item{value}{Credential value to verify.}
}
\description{
Helper function for getting valid credentials
}
