% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{delete_asset}
\alias{delete_asset}
\title{Delete an asset}
\usage{
delete_asset(asset_uuid, use_proxy = FALSE)
}
\arguments{
\item{asset_uuid}{UUID of asset}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Delete an asset
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/deleteAsset}{API documentation}.

Note: This functionality is disabled in Scottish Government workspaces.
}
