% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file_version}
\alias{download_file_version}
\title{Download a document version and save to disk}
\usage{
download_file_version(
  document_uuid,
  folder,
  overwrite = FALSE,
  use_proxy = FALSE
)
}
\arguments{
\item{document_uuid}{UUID of existing document version}

\item{folder}{Folder to save downloaded file to}

\item{overwrite}{Logical to indicate whether file should be overwritten if
already exists.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
An httr2 \link[=response]{httr2::response()} (invisibly)
}
\description{
Download a document version and save to disk
}
