% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveR.R
\name{error}
\alias{error}
\title{Translate error code into helpful message}
\usage{
error(response)
}
\arguments{
\item{response}{An httr2 \link[=response]{httr2::response()}}
}
\value{
A character vector to include in error message.
}
\description{
Translate error code into helpful message
}
