% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{get_proxy}
\alias{get_proxy}
\title{Get OBJECTIVER_PROXY environment variable}
\usage{
get_proxy()
}
\value{
Value of OBJECTIVER_PROXY environment variable
}
\description{
Get OBJECTIVER_PROXY environment variable
}
\examples{
\dontrun{get_proxy()}

}
