% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_id.R
\name{get_user_id}
\alias{get_user_id}
\title{Get user ID on current authenticated user}
\usage{
get_user_id(use_proxy = FALSE)
}
\arguments{
\item{use_proxy}{Logical; to indicate whether proxy required.}
}
\value{
A character value of the user ID
}
\description{
Get user ID on current authenticated user
}
