% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{input_value}
\alias{input_value}
\title{Look for or ask for input value}
\usage{
input_value(type = c("usr", "pwd", "proxy"))
}
\arguments{
\item{type}{One of 'usr' (username), 'pwd' (password), or 'proxy'.}
}
\value{
Validated value
}
\description{
The function looks for the relevant value from environment
variables. If this doesn't exist and the session is interactive, the user
will be prompted to input a value. The value is then validated using
\code{check_valid()}.

Expected environment variables for each \code{type} are as follows:
\itemize{
\item usr: \code{OBJECTIVER_USR}
\item pwd: \code{OBJECTIVER_PWD}
\item proxy: \code{OBJECTIVER_PROXY}
}
}
