% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_auth.R
\name{mobile_auth_login}
\alias{mobile_auth_login}
\title{Login using mobile authenticator}
\usage{
mobile_auth_login(code = NULL, use_proxy = FALSE)
}
\arguments{
\item{code}{Character string. Time-based one time-password from mobile
authenticator. If not supplied, a pop-up window will prompt the user to
enter.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
Login using mobile authenticator
}
\details{
More information on mobile authentication can be found in
\code{vignette}.

More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/MobileAuth/login}{API documentation}.
}
