% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_auth.R
\name{mobile_auth_status}
\alias{mobile_auth_status}
\title{Get mobile authenticator details of the current authenticated user}
\usage{
mobile_auth_status(use_proxy = FALSE)
}
\arguments{
\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
A list object containing 2 logical values:
\itemize{
\item \code{mobileAuthLogin}: Has the user enabled login via Mobile Authenticator?
\item \code{mobileAuthRegistered}: Has the user registered a Mobile Authenticator?
}
}
\description{
Get mobile authenticator details of the current authenticated user
}
\details{
More information on mobile authentication can be found in
\code{vignette("authentication")}.

More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/MobileAuth/getMobileAuthDetails}{API documentation}.
}
