% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_user_id.R
\name{my_user_id}
\alias{my_user_id}
\title{Get user ID for current authenticated user}
\usage{
my_user_id(use_proxy = FALSE)
}
\arguments{
\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
A character value of the authenticated user's ID
}
\description{
Get user ID for current authenticated user
}
\examples{
\dontrun{
my_user_id()
}

}
