% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{new_document_version}
\alias{new_document_version}
\title{Create a new document version}
\usage{
new_document_version(file, document_uuid, use_proxy = FALSE)
}
\arguments{
\item{file}{File path of document to upload}

\item{document_uuid}{UUID of existing document}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Create a new document version
}
