% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{new_reply}
\alias{new_reply}
\title{Create a new reply to a thread}
\usage{
new_reply(thread_uuid, text, use_proxy = FALSE)
}
\arguments{
\item{thread_uuid}{UUID of thread to reply to}

\item{text}{Character string to include in body of thread}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
API response (invisibly)
}
\description{
Create a new reply to a thread
}
