% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{new_task}
\alias{new_task}
\title{Create a new task in a workspace}
\usage{
new_task(
  title,
  type = c("acknowledge", "approve", "review"),
  workspace_uuid,
  assignee_uuids,
  description = NULL,
  due_date = NULL,
  asset_uuids = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{title}{String. Task title.}

\item{type}{One of "acknowledge", "approve", "review".}

\item{workspace_uuid}{UUID of workspace to create task in.}

\item{assignee_uuids}{Optionally, a character vector of participant UUIDs
to assign task to.}

\item{description}{Optional string task description.}

\item{due_date}{Optionally, a \code{POSIXct} value to set due date of task. If no
time is supplied, midnight will be used. For example, if \code{"2025-01-01"} is
supplied, the due date will be \code{"2025-01-01 00:00:00 GMT"}}

\item{asset_uuids}{Optionally, a character vector of asset UUIDs to link the
task to.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
UUID of new task (invisibly)
}
\description{
Create a new task in a workspace
}
