% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{new_thread}
\alias{new_thread}
\title{Create a new thread}
\usage{
new_thread(
  workspace_uuid,
  text,
  mentioned_assets = NULL,
  mentioned_users = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{workspace_uuid}{UUID of workspace}

\item{text}{Character string to include in body of thread}

\item{mentioned_assets}{UUID(s) of asset(s) to mention}

\item{mentioned_users}{UUID(s) of user(s) to mention}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
Create a new thread
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Comments/createThread}{API documentation}.
}
\seealso{
Other Comment functions: 
\code{\link{comments}()},
\code{\link{new_reply}()}
}
\concept{Comment functions}
