% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{new_version}
\alias{new_version}
\title{Upload a file to create a new document version}
\usage{
new_version(file, document_uuid, use_proxy = FALSE)
}
\arguments{
\item{file}{File path of document to upload}

\item{document_uuid}{UUID of existing document}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Upload a file to create a new document version
}
