% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveR.R
\name{objectiveR}
\alias{objectiveR}
\title{Core request function}
\usage{
objectiveR(
  endpoint,
  url_path = NULL,
  url_query = NULL,
  method = "GET",
  body = NULL,
  accept = "application/json",
  content_type = "application/json",
  use_proxy = FALSE
)
}
\arguments{
\item{endpoint}{The endpoint to append to the API server address}

\item{method}{HTTP method to use; e.g. \code{GET}, \code{POST}, \code{PUT}.
Defaults to \code{GET}.}

\item{accept}{Accept header}

\item{content_type}{Content-Type header}

\item{use_proxy}{Logical to indicate whether to use proxy}

\item{...}{Named parameters to pass to request body}
}
\value{
An httr2 \link[=response]{httr2::response()}
}
\description{
Core request function
}
