% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objr-package.R
\docType{package}
\name{objr-package}
\alias{objr-package}
\title{objr: Wrapper for Objective APIs}
\description{
An R package wrapper for Objective APIs including Objective Connect.
}
\references{
\itemize{
\item \href{https://ScotGovAnalysis.github.io/objr/}{\{objr\} package website}
\item \href{https://scotgovanalysis.github.io/objr/articles/authentication.html}{API authentication}
\item \href{https://scotgovanalysis.github.io/objr/articles/two-factor.html}{Two-factor authentication}
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/}{Objective Connect API documentation}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ScotGovAnalysis/objr}
  \item \url{https://ScotGovAnalysis.github.io/objr/}
  \item Report bugs at \url{https://github.com/ScotGovAnalysis/objr/issues}
}

}
\author{
\strong{Maintainer}: Alice Hannah \email{alice.hannah@gov.scot}

Other contributors:
\itemize{
  \item Scottish Government \email{statistics.enquiries@gov.scot} [copyright holder]
}

}
\keyword{internal}
