% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objr.R
\name{objr}
\alias{objr}
\title{Core request function}
\usage{
objr(
  endpoint,
  url_path = NULL,
  url_query = NULL,
  method = "GET",
  body = NULL,
  path = NULL,
  accept = "application/json",
  content_type = "application/json",
  use_proxy = FALSE
)
}
\arguments{
\item{endpoint}{The endpoint to append to the API server address}

\item{url_path}{A list of values to be added to the request URL path.
Values will be separated with \code{/}.}

\item{url_query}{A list of named values to define query parameters}

\item{method}{HTTP method to use; e.g. \code{GET}, \code{POST}, \code{PUT}.
Defaults to \code{GET}.}

\item{body}{A list of named values to be passed to the request body}

\item{path}{Optional file path to save body of request (mainly used when
downloading files)}

\item{accept}{Accept header. Defaults to 'application/json'.}

\item{content_type}{Content-Type header. Defaults to 'application/json'.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
API response
}
\description{
Core request function
}
\details{
More details on endpoints are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/}{API documentation}.
}
