% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objr.R
\name{objr_auth}
\alias{objr_auth}
\title{Authenticate HTTP request}
\usage{
objr_auth(req)
}
\arguments{
\item{req}{An \href{https://httr2.r-lib.org/reference/request.html}{httr2 request}}
}
\value{
A modified \href{https://httr2.r-lib.org/reference/request.html}{httr2 request}
}
\description{
This sets the authorisation header of an HTTP request. If a
token variable exists in global environment, this is used, otherwise,
the user is prompted to enter an authenticating username and password.
}
\examples{
\dontrun{
httr2::request("http://example.com") |>
  objr_auth()
}

token <- "test"
httr2::request("http://example.com") |> objr_auth()

}
