% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass_2fa.R
\name{participant_bypass_2fa}
\alias{participant_bypass_2fa}
\title{Allow/disallow bypassing of two factor authentication for participant}
\usage{
participant_bypass_2fa(
  participant_uuid,
  allow_bypass = TRUE,
  use_proxy = FALSE
)
}
\arguments{
\item{participant_uuid}{Participant UUID (note that this is different to the
user UUID)}

\item{allow_bypass}{Logical to indicate whether the participant should be
able to bypass two step verification for workspace.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Allow/disallow bypassing of two factor authentication for participant
}
\details{
Note: This setting can only be updated by a workspace owner.
}
