% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{participants}
\alias{participants}
\title{Get data frame of workspace participants}
\usage{
participants(workspace_uuid, use_proxy = FALSE)
}
\arguments{
\item{workspace_uuid}{UUID of workspace}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Data frame
}
\description{
Get data frame of workspace participants
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Participants/getParticipant}{API documentation}.
}
