% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_data}
\alias{read_data}
\title{Read a data file into R}
\usage{
read_data(document_uuid, ..., use_proxy = FALSE)
}
\arguments{
\item{document_uuid}{UUID of existing document}

\item{...}{Additional arguments to pass to read function. See details.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
For csv and xlsx files, a data frame. For rds files, an R object.
}
\description{
Read a data file into R
}
\details{
This function can be used to read the following data file types:
csv, rds, xlsx.

The function works by downloading the file from Objective Connect to a
temporary file and reading it into R. The following functions are used to
read the data and any additional arguments (\code{...}) will be passed to these.\tabular{ll}{
   File Type \tab Function \cr
   csv \tab \code{readr::read_csv()} \cr
   rds \tab \code{readr::read_rds()} \cr
   xlsx \tab \code{readxl::read_xlsx()} \cr
}


To check what file type your document is (and thus what function additional
arguments will be passed to), use \code{asset_info()}.

If there are other data file types you would like to download using this
function, please \href{https://github.com/ScotGovAnalysis/objr/issues/new}{open an issue on the GitHub repository}.
}
