% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{rollback_to_version}
\alias{rollback_to_version}
\title{Rollback a document to a previous version}
\usage{
rollback_to_version(document_uuid, version_uuid, use_proxy = FALSE)
}
\arguments{
\item{document_uuid}{UUID of document (asset)}

\item{version_uuid}{UUID of version to rollback to}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
API response (invisibly)
}
\description{
Rollback a document to a previous version
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/rollbackDocument}{API documentation}.
}
