% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{set_proxy}
\alias{set_proxy}
\title{Set OBJECTIVER_PROXY environment variable}
\usage{
set_proxy(proxy_url = NULL, overwrite = FALSE)
}
\arguments{
\item{proxy_url}{Character value. If NULL, user will be prompted to enter.}

\item{overwrite}{Logical to indicate whether to overwrite variable if it
already exists.}
}
\description{
Set OBJECTIVER_PROXY environment variable
}
\examples{
\dontrun{set_proxy()}

}
