% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_bearer.R
\name{store_bearer}
\alias{store_bearer}
\title{Store bearer token from API response}
\usage{
store_bearer(response, store_env = .GlobalEnv)
}
\arguments{
\item{response}{An httr2 response object; must contain an 'Authorization'
header.}

\item{store_env}{The environment to bind the bearer token to. Defaults to
the global environment.}
}
\value{
Returns the bearer token invisibly. This function is primarily used
for its side effect - an environment variable is created called "bearer".
}
\description{
Store bearer token from API response
}
