% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_token.R
\name{store_token}
\alias{store_token}
\title{Store session token from API response}
\usage{
store_token(response, store_env = globalenv())
}
\arguments{
\item{response}{An httr2 response object; must contain an 'Authorization'
header.}

\item{store_env}{The environment to bind the token to.}
}
\value{
Returns the token invisibly. This function is primarily used
for its side effect - an environment variable is created called "token".
}
\description{
Store session token from API response
}
