% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tasks}
\alias{tasks}
\title{Get tasks in workspaces}
\usage{
tasks(
  workspace_uuid = NULL,
  workspace_owner_uuid = NULL,
  asset_uuid = NULL,
  assigned_to_me = FALSE,
  created_by_me = FALSE,
  status = c("all", "open", "cancelled", "complete"),
  workgroup_uuid = NULL,
  page = NULL,
  size = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{workspace_uuid}{UUID of workspace to filter on}

\item{workspace_owner_uuid}{UUID of workspace owner of workspaces to filter
on}

\item{asset_uuid}{UUID of asset to filter on}

\item{assigned_to_me}{Logical; If \code{TRUE}, will filter on tasks assigned to
authenticated user}

\item{created_by_me}{Logical; If \code{TRUE}, will filter on tasks created by
authenticated user}

\item{status}{One of "all", "open", "cancelled", "complete"}

\item{workgroup_uuid}{UUID of workgroup to filter on}

\item{page}{Page number of responses to return (0..N).}

\item{size}{Number of results to be returned per page.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Tibble
}
\description{
Get tasks in workspaces
}
\examples{
\dontrun{
tasks()
}

}
