% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file to create a new document}
\usage{
upload_file(
  file,
  workspace_uuid,
  name = NULL,
  description = NULL,
  parent_uuid = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{file}{File path of document to upload}

\item{workspace_uuid}{UUID of the workspace to create the new document in}

\item{name}{Name to give document. If this isn't provided, the name of the
file will be used.}

\item{description}{Optional description of document.}

\item{parent_uuid}{UUID of folder in the workspace to create the new
document within. If not supplied, the document will be created in the
top-level of the workspace.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
An httr2 \link[=response]{httr2::response()} (invisibly)
}
\description{
Upload a file to create a new document
}
