% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass_2fa.R
\name{user_bypass_2fa}
\alias{user_bypass_2fa}
\title{Allow/disallow bypassing of two factor authentication for user}
\usage{
user_bypass_2fa(user_uuid, allow_bypass = TRUE, use_proxy = FALSE)
}
\arguments{
\item{user_uuid}{Workgroup UUID}

\item{allow_bypass}{Logical to indicate whether the user should be able
to bypass two step verification for workspace.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Allow/disallow bypassing of two factor authentication for user
}
\details{
Note: This setting can only be updated by a workspace owner.
}
