% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_input}
\alias{valid_input}
\title{Helper function for getting valid input value}
\usage{
valid_input(type = c("usr", "pwd", "proxy"), value = NULL)
}
\arguments{
\item{type}{One of 'usr' (username), 'pwd' (password), or 'proxy'.}

\item{value}{Value to verify. If NULL, user will be prompted to enter.}
}
\description{
Helper function for getting valid input value
}
