% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{versions}
\alias{versions}
\title{Get data frame of document versions}
\usage{
versions(document_uuid, page = NULL, size = NULL, use_proxy = FALSE)
}
\arguments{
\item{document_uuid}{UUID of document (asset)}

\item{page}{Page number of responses to return (0..N). Default is 0.}

\item{size}{Number of results to be returned per page. Default is 20.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Data frame
}
\description{
Get data frame of document versions
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/getVersionsByDocumentUuid}{API documentation}.
}
