% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass_2fa.R
\name{workgroup_bypass_2fa}
\alias{workgroup_bypass_2fa}
\title{Allow/disallow bypassing of two-factor authentication for workgroup}
\usage{
workgroup_bypass_2fa(workgroup_uuid, allow_bypass = TRUE, use_proxy = FALSE)
}
\arguments{
\item{workgroup_uuid}{Workgroup UUID}

\item{allow_bypass}{Logical to indicate whether the workgroup should allow
selected participants to bypass two-factor authentication verification.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
Allow/disallow bypassing of two-factor authentication for workgroup
}
\details{
More information on two-factor authentication can be found in
\code{vignette("two-factor")}.

More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Workgroups/setWorkgroupMfaBypassAllowed}{API documentation}.
}
