% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass_2fa.R
\name{workgroup_mandate_2fa}
\alias{workgroup_mandate_2fa}
\title{Enable/disable mandatory two-factor authentication for workgroup}
\usage{
workgroup_mandate_2fa(workgroup_uuid, mandate = TRUE, use_proxy = FALSE)
}
\arguments{
\item{workgroup_uuid}{Workgroup UUID}

\item{mandate}{Logical to indicate whether two-factor authentication should
be mandatory in the workgroup}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
Enable/disable mandatory two-factor authentication for workgroup
}
\details{
More information on two-factor authentication can be found in
\code{vignette("two-factor")}.

More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html#/Workgroups/setTwoStepMandatory}{API documentation}.
}
\seealso{
Other Two-factor authentication functions: 
\code{\link{participant_bypass_2fa}()},
\code{\link{workgroup_bypass_2fa}()}
}
\concept{Two-factor authentication functions}
