% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{workspaces}
\alias{workspaces}
\title{Get workspaces the current user is a member of}
\usage{
workspaces(workgroup_uuid = NULL, page = NULL, size = NULL, use_proxy = FALSE)
}
\arguments{
\item{workgroup_uuid}{UUID of workgroup to filter by}

\item{page}{Page number of responses to return (0..N).}

\item{size}{Number of results to be returned per page.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Data frame
}
\description{
Get workspaces the current user is a member of
}
