% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{write_data}
\alias{write_data}
\title{Write a data file from R to create a new document}
\usage{
write_data(
  x,
  file_name,
  file_type,
  workspace_uuid,
  ...,
  description = NULL,
  parent_uuid = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{x}{R object to write to file.}

\item{file_name}{Name to give file.}

\item{file_type}{Either "csv", "rds" or "xlsx".}

\item{workspace_uuid}{UUID of the workspace to create the new document in}

\item{...}{Additional arguments to pass to write function. See details.}

\item{description}{Optional description of document.}

\item{parent_uuid}{UUID of folder in the workspace to create the new
document within. If not supplied, the document will be created in the
top-level of the workspace.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
An httr2 \link[=response]{httr2::response()} (invisibly)
}
\description{
Write a data file from R to create a new document
}
\details{
This function can be used to write the following data file types:
csv, rds, xlsx. Use the \code{file_type} argument to control which file type
to create.

The function works by writing the R object to a temporary file and uploading
the file to Objective Connect. The following functions are used to
write the data and any additional arguments (\code{...}) will be passed to these.\tabular{ll}{
   File Type \tab Function \cr
   csv \tab \code{readr::write_csv()} \cr
   rds \tab \code{readr::write_rds()} \cr
   xlsx \tab \code{writexl::write_xlsx()} \cr
}


If there are other data file types you would like to upload using this
function, please \href{https://github.com/ScotGovAnalysis/objr/issues/new}{open an issue on the GitHub repository}.
}
