% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{asset_info}
\alias{asset_info}
\title{Get asset information}
\usage{
asset_info(asset_uuid, use_proxy = FALSE)
}
\arguments{
\item{asset_uuid}{UUID of asset}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
A tibble
}
\description{
Get asset information
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/getAssetByUuid}{API documentation}.
}
