% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets}
\alias{assets}
\title{Get data frame of assets in workspace}
\usage{
assets(
  workspace_uuid,
  type = NULL,
  page = NULL,
  size = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{workspace_uuid}{UUID of workspace}

\item{type}{Asset type to filter results by. Either "document", "folder" or
"link". Default returns all asset types.}

\item{page}{Page number of responses to return (0..N). Default is 0.}

\item{size}{Number of results to be returned per page. Default is 20.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
A tibble
}
\description{
Get data frame of assets in workspace
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/getAssets}{API documentation}.
}
\seealso{
Other Asset functions: 
\code{\link{asset_info}()},
\code{\link{create_folder}()},
\code{\link{delete_asset}()},
\code{\link{rename_asset}()}
}
\concept{Asset functions}
