% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets}
\alias{assets}
\title{Get data frame of assets in workspace}
\usage{
assets(
  workspace_uuid,
  type = list("document", "folder", "link"),
  page = NULL,
  size = NULL,
  use_proxy = FALSE
)
}
\arguments{
\item{workspace_uuid}{UUID of workspace}

\item{type}{List of asset types to return. Defaults to all types;
document, folder and link.}

\item{page}{Page number of responses to return (0..N).}

\item{size}{Number of results to be returned per page.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Tibble
}
\description{
Get data frame of assets in workspace
}
