% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{delete_asset}
\alias{delete_asset}
\title{Delete an asset}
\usage{
delete_asset(asset_uuid, use_proxy = FALSE)
}
\arguments{
\item{asset_uuid}{UUID of asset}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Delete an asset
}
\details{
Note: Note: This functionality is disabled in Scottish Government
workspaces.
}
