% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\alias{download_file_version}
\title{Download a file and save to disk}
\usage{
download_file(
  document_uuid,
  folder,
  file_name = NULL,
  overwrite = FALSE,
  use_proxy = FALSE
)

download_file_version(
  document_uuid,
  folder,
  file_name = NULL,
  overwrite = FALSE,
  use_proxy = FALSE
)
}
\arguments{
\item{document_uuid}{UUID of asset or document version}

\item{folder}{Folder to save downloaded file to}

\item{file_name}{Optional name to give downloaded file. If not provided,
file will have same name as Objective Connect asset.}

\item{overwrite}{Logical to indicate whether file should be overwritten if
already exists. Defaults to \code{FALSE}.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Path to downloaded file (invisibly).
}
\description{
\itemize{
\item Use \code{download_file()} with an asset UUID for the latest version of a
document.
\item Use \code{download_file_version()} with a document version UUID for a specific
version of a document.
}
}
\details{
More details on the endpoints used by these functions are available in the
API documentation:
\itemize{
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/download}{\code{download_file}}
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/downloadVersion}{\code{download_file_version}}
}
}
