% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_file}
\alias{download_file}
\title{Download a file}
\usage{
download_file(document_uuid, folder, use_proxy = FALSE)
}
\arguments{
\item{document_uuid}{UUID of existing document}

\item{folder}{Folder to save downloaded file to}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\description{
Download a file
}
