% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_user_id.R
\name{my_user_id}
\alias{my_user_id}
\title{Get user ID for current authenticated user}
\usage{
my_user_id(use_proxy = FALSE)
}
\arguments{
\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
A character value of the authenticated user's ID
}
\description{
Get user ID for current authenticated user
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Users/getCurrentUser}{API documentation}.
}
\examples{
\dontrun{
my_user_id()
}

}
