% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveR_auth.R
\name{objectiveR_auth}
\alias{objectiveR_auth}
\title{Authenticate HTTP request}
\usage{
objectiveR_auth(req)
}
\arguments{
\item{req}{An httr2 \link[=request]{httr2::request()}}
}
\value{
A modified httr2 \link[=request]{httr2::request()}
}
\description{
This sets the authorisation header of an HTTP request. If a
token variable exists in global environment, this is used, otherwise,
the user is prompted to enter an authenticating username and password.
}
\examples{
\dontrun{
httr2::request("http://example.com") |>
  objectiveR_auth()
}

token <- "test"
httr2::request("http://example.com") |> objectiveR_auth()

}
