% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectiveR_auth.R
\name{objectiveR_auth}
\alias{objectiveR_auth}
\title{Authenticate HTTP request}
\usage{
objectiveR_auth(req, usr = NULL, pwd = NULL)
}
\arguments{
\item{req}{An httr2 \link[=request]{httr2::request()}}

\item{usr, pwd}{Username and password for basic authentication.
These arguments are required when using basic authentication in a
non-interactive environment. If working interactively, these
can be ignored and RStudio will prompt you to enter.}
}
\value{
A modified httr2 \link[=request]{httr2::request()}
}
\description{
This sets the authorisation header of an HTTP request. If a
token variable exists in global environment, this is used, otherwise,
the user is prompted to enter an authenticating username and password.
}
\examples{
req <-
  httr2::request("http://example.com") |>
  objectiveR_auth(usr = "ex-usr", pwd = "ex-pwd")
req
req |> httr2::req_dry_run()

token <- "test"
req <- httr2::request("http://example.com") |> objectiveR_auth()
req
req |> httr2::req_dry_run()

}
