% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objr-package.R
\docType{package}
\name{objr-package}
\alias{objr-package}
\title{objr: Wrapper for Objective APIs}
\description{
An R package wrapper for Objective APIs including Objective Connect.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ScotGovAnalysis/objr}
  \item \url{https://ScotGovAnalysis.github.io/objr/}
  \item Report bugs at \url{https://github.com/ScotGovAnalysis/objr/issues}
}

}
\author{
\strong{Maintainer}: Alice Hannah \email{alice.hannah@gov.scot}

Other contributors:
\itemize{
  \item Scottish Government \email{statistics.enquiries@gov.scot} [copyright holder]
}

}
\keyword{internal}
