% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_data}
\alias{read_data}
\alias{read_data_version}
\title{Read a data file into R}
\usage{
read_data(document_uuid, ..., use_proxy = FALSE)

read_data_version(document_uuid, ..., use_proxy = FALSE)
}
\arguments{
\item{document_uuid}{UUID of asset or document version}

\item{...}{Additional arguments passed to read function. See details.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\itemize{
\item Use \code{read_data()} with an asset UUID for the latest version of a document.
\item Use \code{read_data_version()} with a document version UUID for a specific
version of a document.
}
}
\details{
This function can be used to read the following data file types:
csv, rds, xlsx.

The function works by downloading the file from Objective Connect to a
temporary file and reading it into R. The following functions are used to
read the data and any additional arguments (\code{...}) will be passed to these.\tabular{ll}{
   File Type \tab Function \cr
   csv \tab \code{readr::read_csv()} \cr
   rds \tab \code{readr::read_rds()} \cr
   xlsx \tab \code{readxl::read_xlsx()} \cr
}


To check what file type your document is (and thus what function additional
arguments will be passed to), use \code{asset_info()}.

If there are other data file types you would like to download using this
function, please \href{https://github.com/ScotGovAnalysis/objr/issues/new}{open an issue on the GitHub repository}.

More details on the endpoints used by these functions are available in the
API documentation:
\itemize{
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/download}{\code{read_data}}
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/downloadVersion}{\code{read_data_version}}
}
}
