% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{rename_asset}
\alias{rename_asset}
\title{Rename an asset}
\usage{
rename_asset(asset_uuid, new_name, use_proxy = FALSE)
}
\arguments{
\item{asset_uuid}{UUID of asset}

\item{new_name}{Character. New name to give asset.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
Rename an asset
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/updateAssetName}{API documentation}.
}
