% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_file}
\alias{upload_file}
\alias{upload_file_version}
\title{Upload a file}
\usage{
upload_file(
  file,
  uuid,
  name = NULL,
  description = NULL,
  parent_uuid = NULL,
  use_proxy = FALSE
)

upload_file_version(file, uuid, use_proxy = FALSE)
}
\arguments{
\item{file}{File path of document to upload}

\item{uuid}{For \code{upload_file()}, a workspace UUID to create the new document
in. For \code{upload_file_version()}, an asset UUID to create a new version of.}

\item{name}{Name to give document. If this isn't provided, the name of the
file will be used.}

\item{description}{Optional description of document.}

\item{parent_uuid}{UUID of folder in the workspace to create the new
document within. If not supplied, the document will be created in the
top-level of the workspace.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
\itemize{
\item Use \code{upload_file()} to create a new file in a workspace.
\item Use \code{upload_file_version()} to create a new version of an existing
document.
}
}
\details{
More details on the endpoints used by these functions are available in the
API documentation:
\itemize{
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/createDocument}{\code{upload_file}}
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/createDocumentVersion_1}{\code{upload_file_version}}
}
}
