% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{workspaces}
\alias{workspaces}
\title{Get workspaces the current user is a member of}
\usage{
workspaces(workgroup_uuid = NULL, page = NULL, size = NULL, use_proxy = FALSE)
}
\arguments{
\item{workgroup_uuid}{UUID of workgroup to filter by}

\item{page}{Page number of responses to return (0..N). Default is 0.}

\item{size}{Number of results to be returned per page. Default is 20.}

\item{use_proxy}{Logical to indicate whether to use proxy}
}
\value{
Data frame
}
\description{
Get workspaces the current user is a member of
}
\details{
More details on this endpoint are available in the
\href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Workspaces/getMyWorkspaces}{API documentation}.
}
