% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{write_data}
\alias{write_data}
\alias{write_data_version}
\title{Write a data file from R}
\usage{
write_data(
  x,
  uuid,
  file_name,
  file_type,
  ...,
  description = NULL,
  parent_uuid = NULL,
  use_proxy = FALSE
)

write_data_version(x, uuid, ..., use_proxy = FALSE)
}
\arguments{
\item{x}{R object to write to file.}

\item{uuid}{Either a workspace UUID to create a new document or an asset UUID
to create a new version of an existing document.}

\item{file_name}{Name to give file.}

\item{file_type}{Either "csv", "rds" or "xlsx".}

\item{...}{Additional arguments to pass to write function. See details.}

\item{description}{Optional description of document.}

\item{parent_uuid}{UUID of folder in the workspace to create the new
document within. If not supplied, the document will be created in the
top-level of the workspace.}

\item{use_proxy}{Logical to indicate whether to use proxy.}
}
\value{
API response (invisibly)
}
\description{
\itemize{
\item Use \code{write_data()} to create a new file in a workspace.
\item Use \code{write_data_version()} to write a data file as a new version of an
existing document.
}
}
\details{
This function can be used to write the following data file types:
csv, rds, xlsx. If writing to a new document, use the \code{file_type}
argument to control which file type to create. If writing a new version of an
existing document, the existing file type will be used.

The function works by writing the R object to a temporary file and uploading
the file to Objective Connect. The following functions are used to
write the data and any additional arguments (\code{...}) will be passed to these.\tabular{ll}{
   File Type \tab Function \cr
   csv \tab \code{readr::write_csv()} \cr
   rds \tab \code{readr::write_rds()} \cr
   xlsx \tab \code{writexl::write_xlsx()} \cr
}


If there are other data file types you would like to upload using this
function, please \href{https://github.com/ScotGovAnalysis/objr/issues/new}{open an issue on the GitHub repository}.

More details on the endpoints used by these functions are available in the
API documentation:
\itemize{
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/createDocument}{\code{write_data}}
\item \href{https://secure.objectiveconnect.co.uk/publicapi/1/swagger-ui/index.html?configUrl=/publicapi/1/v3/api-docs/swagger-config#/Assets/createDocumentVersion_1}{\code{write_data_version}}
}
}
